Add files via upload
This is a web dev beginner project to create a weather website and fetch the weather API using HTML, CSS and JavaScript. It consists of a standard card with a search bar and button which displays the temperature, weather, humidity and windspeed along along with a navbar which consists of source which takes user to the official website of OpenWeather and About Us goes to my github page. The background image loads itself according to the weather fetched by the web.